편지과제

+++ src/main/java/org/yejin/article/ArticleDao.java
@@ -19,9 +19,9 @@ public class ArticleDao {
 
 	static final String ADD_ARTICLE = "insert article(title,content,userId,name) values(?,?,?,?)";
 
-	static final String EDIT_ARTICLE = "update article set title=?, content=? where articleId=?";
+	static final String UPDATE_ARTICLE = "update article set title=?, content=? where (articleId, userId) = (?,?)";
 	
-	static final String DELETE_ARTICLE = "delete from article where articleId=?";
+	static final String DELETE_ARTICLE = "delete from article where (articleId, userId) = (?,?)";
 	
 	@Autowired
 	JdbcTemplate jdbcTemplate;
@@ -63,15 +63,15 @@ public class ArticleDao {
 	/**
 	 * 글 수정
 	 */
-	public int editArticle(Article article) {
-		return jdbcTemplate.update(EDIT_ARTICLE, article.getTitle(),
-				article.getContent(), article.getArticleId());
+	public int updateArticle(Article article) {
+		return jdbcTemplate.update(UPDATE_ARTICLE, article.getTitle(),
+				article.getContent(), article.getArticleId(), article.getUserId());
 	}
 	
 	/**
 	 * 글 삭제
 	 */
-	public int deleteArticle(String articleId) {
-		return jdbcTemplate.update(DELETE_ARTICLE, articleId);
+	public int deleteArticle(String articleId, String userId) {
+		return jdbcTemplate.update(DELETE_ARTICLE, articleId, userId);
 	}
 }
